refactor: move selectedConsumer to partitions
